New links in Draft View
